+++ RELEASES.md
@@ -1,7 +1,11 @@
 ## Version 5.10.1
 Added a JSONInsance class to allowing saving/load of state machine state as a JSON string
 
-Deprecate FinalStateClass: as it is just a State that has no outbound transitions.
+Acknowledge that IInstance.getCurrent could return undefined if the Region had not previously been entered.
+
+Replace a few ```let``` declarations with ```const``` where the value is not changed.
+
+Deprecate FinalStateClass as it is just a State that has no outbound transitions; this will be droped in the 6.0 release.
 
 ## Version 5.10.0
 Moved to TypeScript 2.0: